Refactored the fieldtypes.py into a folder containing the fieldtypes per file + changed imports where necessary to make sure code and tests still work
