+++ api/app/signals/apps/questionnaires/fieldtypes/integer.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2022 Gemeente Amsterdam
+from signals.apps.questionnaires.fieldtypes.base import FieldType
+
+
+class Integer(FieldType):
+    submission_schema = {'type': 'integer'}
