+++ api/app/signals/apps/questionnaires/services/answer.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: MPL-2.0
-# Copyright (C) 2021 Gemeente Amsterdam
+# Copyright (C) 2021 - 2022 Gemeente Amsterdam
 from django.core.exceptions import ValidationError as django_validation_error
 
 from signals.apps.questionnaires.fieldtypes import get_field_type_class
